Moved the private types to __init__ and the files they are used.
